DNS searching function acl_dns_lookup() in acl_dns.c can send query to multiple DNS and use the first reached result.

+++ lib_acl/src/net/dns/rfc1035.c
@@ -677,7 +677,7 @@ static int rfc1035RRPack(const rfc1035_rr *RR, char *buf, size_t sz)
 	return (off);
 }
 
-ssize_t rfc1035BuildAReply(const char *hostname, const ACL_ARGV *ip_argv,
+size_t rfc1035BuildAReply(const char *hostname, const ACL_ARGV *ip_argv,
 	const char *domain_root, const char *dnsname, const char *dns_ip,
 	unsigned short qid, char *buf, size_t sz)
 {
@@ -744,7 +744,7 @@ ssize_t rfc1035BuildAReply(const char *hostname, const ACL_ARGV *ip_argv,
 		acl_myfree(rr.rdata);
 	}
 
-	return ((int) offset);
+	return offset;
 }
 /*
 * rfc1035BuildAQuery()
@@ -756,7 +756,7 @@ ssize_t rfc1035BuildAReply(const char *hostname, const ACL_ARGV *ip_argv,
 * the size of the message (i.e. how much to write).
 * Returns the size of the query
 */
-ssize_t rfc1035BuildAQuery(const char *hostname, char *buf, size_t sz,
+size_t rfc1035BuildAQuery(const char *hostname, char *buf, size_t sz,
 	unsigned short qid, rfc1035_query * query)
 {
 	const char *myname = "rfc1035BuildAQuery";
@@ -780,7 +780,7 @@ ssize_t rfc1035BuildAQuery(const char *hostname, char *buf, size_t sz,
 
 	if (offset > sz)
 		acl_msg_fatal("%s: offset(%d) > sz(%d)", myname, (int) offset, (int) sz);
-	return (ssize_t) offset;
+	return offset;
 }
 
 /*
@@ -793,7 +793,7 @@ ssize_t rfc1035BuildAQuery(const char *hostname, char *buf, size_t sz,
 * the size of the message (i.e. how much to write).
 * Returns the size of the query
 */
-ssize_t rfc1035BuildPTRQuery(const struct in_addr addr, char *buf,
+size_t rfc1035BuildPTRQuery(const struct in_addr addr, char *buf,
 	size_t sz, unsigned short qid, rfc1035_query * query)
 {
 	const char *myname = "rfc1035BuildPTRQuery";
@@ -821,7 +821,7 @@ ssize_t rfc1035BuildPTRQuery(const struct in_addr addr, char *buf,
 	}
 	if (offset > sz)
 		acl_msg_fatal("%s: offset(%d) > sz(%d)", myname, (int) offset, (int) sz);
-	return (ssize_t) offset;
+	return offset;
 }
 
 /*
